Refactored the codebase.
