+++ modules/carbon-datasources/carbon-datasources/feature/org.wso2.carbon.datasource.core.feature/pom.xml
@@ -22,18 +22,11 @@
         <dependency>
             <groupId>commons-io.wso2</groupId>
             <artifactId>commons-io</artifactId>
-            <version>2.4.0.wso2v1</version>
         </dependency>
         <dependency>
             <groupId>org.apache.tomcat.wso2</groupId>
             <artifactId>jdbc-pool</artifactId>
-            <version>7.0.34.wso2v2</version>
         </dependency>
-        <!--<dependency>-->
-            <!--<groupId>org.apache.tomcat</groupId>-->
-            <!--<artifactId>tomcat-jdbc</artifactId>-->
-            <!--<version>7.0.59</version>-->
-        <!--</dependency>-->
     </dependencies>
 
     <build>
@@ -73,8 +66,6 @@
                                 <bundle>
                                     <symbolicName>jdbc-pool</symbolicName>
                                     <version>7.0.34.wso2v2</version>
-                                    <!--<symbolicName>org.apache.tomcat.jdbc</symbolicName>-->
-                                    <!--<version>7.0.59</version>-->
                                 </bundle>
                             </bundles>
                         </configuration>
