+++ modules/carbon-datasources/carbon-datasources/org.wso2.carbon.datasource.core/src/main/java/org/wso2/carbon/datasource/rdbms/hikari/HikariDataSourceReader.java
@@ -46,9 +46,4 @@ public class HikariDataSourceReader implements DataSourceReader {
                     e.getMessage(), e);
         }
     }
-
-    @Override
-    public boolean testDataSourceConnection(String xmlConfiguration) throws DataSourceException {
-        return false;
-    }
 }
