+++ modules/carbon-datasources/carbon-datasources/org.wso2.carbon.datasource.core/src/main/resources/master-datasources.xml
@@ -2,7 +2,7 @@
 
     <providers>
         <!--<provider>org.wso2.carbon.datasource.rdbms.hikari.HikariDataSourceReader</provider>-->
-        <provider>org.wso2.carbon.datasource.rdbms.tomcat.RDBMSDataSourceReader</provider>
+        <provider>org.wso2.carbon.datasource.rdbms.tomcat.TomcatSDataSourceReader</provider>
     </providers>
 
     <datasources>
@@ -11,12 +11,14 @@
             <name>WSO2_CARBON_DB</name>
             <description>The datasource used for registry and user manager</description>
             <jndiConfig>
-                <name>jdbc/WSO2CarbonDB/test</name>
-                <useDataSourceFactory>true</useDataSourceFactory>
-                <environment>
-                    <property name="java.naming.factory.initial">com.sun.jndi.fscontext.RefFSContextFactory</property>
-                    <property name="java.naming.provider.url">file:///home/dinushab/Carbon/POC/data-sources/carbon-kernel/modules/carbon-datasources/carbon-datasources/org.wso2.carbon.datasource.core/src/test</property>
-                </environment>
+                <name>jdbc</name>
+                <!-- useDataSourceFactory>true</useDataSourceFactory -->
+                <!--<environment>-->
+                    <!--<property name="java.naming.factory.initial">com.sun.jndi.fscontext.RefFSContextFactory</property>-->
+                    <!--<property name="java.naming.provider.url">-->
+                        <!--file:///home/dinushab/Carbon/POC/data-sources/carbon-kernel/modules/carbon-datasources/carbon-datasources/org.wso2.carbon.datasource.core/src/test-->
+                    <!--</property>-->
+                <!--</environment>-->
             </jndiConfig>
             <definition type="RDBMS">
                 <configuration>
@@ -24,7 +26,7 @@
                     <username>root</username>
                     <password>root</password>
                     <driverClassName>com.mysql.jdbc.Driver</driverClassName>
-                    <!--<driverClassName>org.h2.Driver</driverClassName>-->
+                    <!-- driverClassName>org.h2.Driver</driverClassName -->
                     <maxActive>50</maxActive>
                     <maxWait>60000</maxWait>
                     <testOnBorrow>true</testOnBorrow>
